requirements/prod.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.4

# Templates
Markdown
#django-widget-tweaks
django-crispy-forms

# Security
bleach
django-session-csrf
#cef

# Logging
#raven

# Celery: Message queue
#celery
#django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
#python-memcached

# Admin
django-debug-toolbar
# There are useful panels for debug-toolbar, but have not quite been updated yet for 1.4
#-e git://github.com/playfire/django-debug-toolbar-user-panel#egg=django_debug_toolbar_user_panel-dev
#-e git://github.com/ross/memcache-debug-panel#egg=memcache_toolbar-dev

# Migrations
South

# Various tidbits
#commonware
#-e git://github.com/mozilla/nuggets.git#egg=nuggets
django-treebeard

# Pagination
#django-pagination==1.0.7
